added softdeps to modinfo - bumped APIVERSION
APIVERSION is now 2. modules written for 2 are compatible with 1;
however modules written for 1 need to be updated (by adding softdeps
to modinfo).

softdeps are loaded before the current module is started, when possible.
however, if they are disabled (by setting them to 0 in config
"[autoloads]"), or non-existent, they will be ignored. Additionally, if
they are unloaded at runtime any modules with softdeps on them will
continue to be loaded and are expected to not raise any errors.